# openAPI.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http:\/\/127.0.0.1:8000\/api"
        }
    ],
    "paths": {
...
        "\/v1\/reservations\/{reservation}\/confirm": {
            "post": {
                "operationId": "reservation.confirm",
                "tags": [
                    "Reservation"
                ],
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "required": true,
                        "description": "The reservation ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Reservation confirmed"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "const": 1
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "data",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "const": "Reservation expired"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                }
                                            },
                                            "required": [
                                                "message",
                                                "status"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "const": "Reservation is not pending"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                }
                                            },
                                            "required": [
                                                "message",
                                                "status"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#\/components\/responses\/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/AuthenticationException"
                    }
                }
            }
        },
...
    },
...
}